chore: replace paver quality tests and js commands
