--- .stylelintignore
@@ -1,5 +1,0 @@
-xmodule/css
-common/static/sass/bourbon
-common/static/xmodule/modules/css
-common/test/test-theme
-lms/static/sass/vendor
