+++ package.json
@@ -12,7 +12,26 @@
     "compile-sass-dev": "scripts/compile_sass.py --env=development",
     "watch": "{ npm run watch-webpack& npm run watch-sass& } && sleep infinity",
     "watch-webpack": "npm run webpack-dev -- --watch",
-    "watch-sass": "scripts/watch_sass.sh"
+    "watch-sass": "scripts/watch_sass.sh",
+    "lint": "python scripts/eslint.py",
+    "test": "npm run test-cms && npm run test-lms && npm run test-xmodule && npm run test-common && npm run test-jest",
+    "test-kind-vanilla": "npm run test-cms-vanilla && npm run test-xmodule-vanilla && npm run test-common-vanilla",
+    "test-kind-require": "npm run test-cms-require && npm run test-common-require",
+    "test-kind-webpack": "npm run test-cms-webpack && npm run test-lms-webpack && npm run test-xmodule-webpack",
+    "test-cms": "npm run test-cms-vanilla && npm run test-cms-require",
+    "test-cms-vanilla": "npm run test-suite -- cms/static/karma_cms.conf.js",
+    "test-cms-require": "npm run test-suite -- cms/static/karma_cms_squire.conf.js",
+    "test-cms-webpack": "npm run test-suite -- cms/static/karma_cms_webpack.conf.js",
+    "test-lms": "echo 'WARNING: Webpack JS tests are disabled. No LMS JS tests will be run. See https://github.com/openedx/edx-platform/issues/35956 for details.'",
+    "test-lms-webpack": "npm run test-suite -- lms/static/karma_lms.conf.js",
+    "test-xmodule": "npm run test-xmodule-vanilla",
+    "test-xmodule-vanilla": "npm run test-suite -- xmodule/js/karma_xmodule.conf.js",
+    "test-xmodule-webpack": "npm run test-suite -- xmodule/js/karma_xmodule_webpack.conf.js",
+    "test-common": "npm run test-common-vanilla && npm run test-common-require",
+    "test-common-vanilla": "npm run test-suite -- common/static/karma_common.conf.js",
+    "test-common-require": "npm run test-suite -- common/static/karma_common_requirejs.conf.js",
+    "test-suite": "${NODE_WRAPPER:-xvfb-run --auto-servernum} node --max_old_space_size=4096 node_modules/.bin/karma start --single-run=true --capture-timeout=60000 --browsers=FirefoxNoUpdates",
+    "test-jest": "jest"
   },
   "dependencies": {
     "@babel/core": "7.26.0",
