+++ requirements/edx/base.txt
@@ -504,7 +504,7 @@ edx-opaque-keys[django]==2.11.0
     #   ora2
 edx-organizations==6.13.0
     # via -r requirements/edx/kernel.in
-edx-proctoring==5.0.1
+edx-proctoring==4.18.4
     # via
     #   -r requirements/edx/kernel.in
     #   edx-proctoring-proctortrack
