+++ requirements/edx/development.txt
@@ -795,7 +795,7 @@ edx-organizations==6.13.0
     # via
     #   -r requirements/edx/doc.txt
     #   -r requirements/edx/testing.txt
-edx-proctoring==5.0.1
+edx-proctoring==4.18.4
     # via
     #   -r requirements/edx/doc.txt
     #   -r requirements/edx/testing.txt
